+++ pnet-data-api-java-sample/src/main/java/pnet/data/api/util/CLI.java
@@ -968,7 +968,7 @@ public class CLI
 
             if (!optionalCommand.isPresent())
             {
-                throw new IllegalArgumentException("Command missing");
+                return;
             }
 
             String command = optionalCommand.get().trim();